pyproject.toml
[tool.poetry]
name = "NMIS_Ecopass"
version = "0.1.14"
description = "A package for creating and managing digital product passports"
authors = ["Syed Awais Hassan Munawar <syed.munawar@strath.ac.uk>"]
readme = "README.md"
packages = [{include = "NMIS_Ecopass", from = "src"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0"
pyyaml = ">=6.0"
lxml = ">=4.9"
openpyxl = ">=3.1"
jmespath = ">=1.0"        # JSON querying
python-dateutil = ">=2.8"  # Better datetime parsing
pyld = ">=2.0.3"          # JSON-LD processing
yamale = ">=4.0"          # YAML schema validation
xmltodict = ">=0.13"      # XML to dict conversion

[tool.poetry.group.dev.dependencies]
black = "^22.3.0"
flake8 = "^3.8.4"
mypy = "^1.0.0"
isort = "^5.0.0"
pre-commit = "^2.0.0"
pytest = "^6.2.5"
pytest-cov = "^2.0.0"
pytest-mock = "^3.0.0"



[[tool.poetry.source]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"